refactor: method depth2 제거 및 기능 분리
