+++ src/main/java/com/step3/util/RandomUtil.java
@@ -1,0 +1,11 @@
+package com.step3.util;
+
+import java.util.Random;
+
+public class RandomUtil {
+    private static final int RANDOM_VALUE_BOUND = 10;
+
+    public static int getRandomValue() {
+        return new Random().nextInt(RANDOM_VALUE_BOUND);
+    }
+}
